Updates doc for spi and re-organizes a class
