+++ spi/README.md
@@ -2,6 +2,12 @@
 
 This SPI provides interfaces to implement Resource Sharing and Access Control.
 
+
+## Usage
+
+A plugin defining a resource and aiming to implement access control over that resource must extend ResourceSharingExtension class to register itself
+
+
 ## License
 
 This code is licensed under the Apache 2.0 License.
