+++ spi/src/main/java/org/opensearch/security/spi/resources/ResourceSharingExtension.java
@@ -9,7 +9,7 @@
 package org.opensearch.security.spi.resources;
 
 /**
- * This interface should be implemented by all the plugins that define one or more resources.
+ * This interface should be implemented by all the plugins that define one or more resources and need access control over those resources.
  *
  * @opensearch.experimental
  */
@@ -17,15 +17,19 @@ public interface ResourceSharingExtension {
 
     /**
      * Type of the resource
-     * @return a string containing the type of the resource
+     * @return a string containing the type of the resource. A qualified class name can be supplied here.
      */
     String getResourceType();
 
     /**
-     * The index where resource meta-data is stored
-     * @return the name of the parent index where resource meta-data is stored
+     * The index where resource is stored
+     * @return the name of the parent index where resource is stored
      */
     String getResourceIndex();
 
+    /**
+     * The parser for the resource, which will be used by security plugin to parse the resource
+     * @return the parser for the resource
+     */
     ResourceParser<? extends Resource> getResourceParser();
 }
