+++ src/main/java/org/opensearch/security/OpenSearchSecurityPlugin.java
@@ -147,6 +147,7 @@ import org.opensearch.security.auth.BackendRegistry;
 import org.opensearch.security.common.resources.ResourceAccessHandler;
 import org.opensearch.security.common.resources.ResourceIndexListener;
 import org.opensearch.security.common.resources.ResourcePluginInfo;
+import org.opensearch.security.common.resources.ResourceProvider;
 import org.opensearch.security.common.resources.ResourceSharingConstants;
 import org.opensearch.security.common.resources.ResourceSharingIndexHandler;
 import org.opensearch.security.common.resources.ResourceSharingIndexManagementRepository;
@@ -195,7 +196,6 @@ import org.opensearch.security.setting.OpensearchDynamicSetting;
 import org.opensearch.security.setting.TransportPassiveAuthSetting;
 import org.opensearch.security.spi.resources.Resource;
 import org.opensearch.security.spi.resources.ResourceParser;
-import org.opensearch.security.spi.resources.ResourceProvider;
 import org.opensearch.security.spi.resources.ResourceSharingExtension;
 import org.opensearch.security.ssl.ExternalSecurityKeyStore;
 import org.opensearch.security.ssl.OpenSearchSecureSettingsFactory;
